requirements.txt
kafka-python==2.0.2
pymongo==4.5.0
pyspark==3.3.0
pandas==2.0.0
numpy==1.24.0
faker==18.0.0
python-dotenv==1.0.0
apache-airflow==2.5.0
apache-airflow-providers-mongo==3.1.1
apache-airflow-providers-spark==3.2.0
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.6
